Removing third party folders. It is now up to the cloner to install the dependencies.
Cleaning up clang-tidy warnings as well.

+++ src/components/BoxCollider.cpp
@@ -41,12 +41,12 @@ void BoxCollider::updateBBox()
 	switch (origin_) 
 	{
 	case Alignment::TOP_LEFT:
-		rect_.x = actorPosition.x + offset_.x;
-		rect_.y = actorPosition.y + offset_.y;
+		rect_.x = (int)(actorPosition.x + offset_.x);
+		rect_.y = (int)(actorPosition.y + offset_.y);
 		break;
 	case Alignment::CENTER_ORIGIN:
-		rect_.x = actorPosition.x - (rect_.w / 2) + offset_.x;
-		rect_.y = actorPosition.y - (rect_.h / 2) + offset_.y;
+		rect_.x = (int)(actorPosition.x - (rect_.w / 2) + offset_.x);
+		rect_.y = (int)(actorPosition.y - (rect_.h / 2) + offset_.y);
 		break;
 	}
